Add type-checker compatible lazy import mode
The --lazy_loader flag enables an inegration with the lazy_loader
library, using it to handle all __init__.py. By default, this library is
incompatible with type checkers such as mypy and pyright. As a
workaround, lazy_loader offers a
[workaround](https://scientific-python.org/specs/spec-0001/#type-checkers)
using `__init__.pyi` stub files. Files to to be imported are read from
the stub file.

Here, add a new --lazy_loader_typed modes that enables this
type-compatible integration. When activated, the vanilla mkinit output
is written into `__init__.pyi` instead of `__init__.py`. The lazy_loader
boilerplate to enable the integration is written into `__init__.py`.